cleanup some todos and move stuff around, update version to 0.7
