+++ gradle.properties
@@ -14,7 +14,8 @@ catalystcore_version=2.1.4
 catalysteffects_version=2.0.1
 btwaila_version=1.2.2-7.3_04
 uselessnumerical_version=1.0.2
+
 # Mod
-mod_version=0.6.0+7.3_04
+mod_version=0.7.0+7.3_04
 mod_group=teamport
 mod_name=better-with-aether
