+++ src/main/java/teamport/aether/items/accessory/SlotAccessory.java
@@ -28,7 +28,6 @@ public class SlotAccessory extends Slot {
             "aether:item/armor_thin_wildcard_outline",
     };
 
-    // TODO add inventory equipment achievements
     public final MenuInventory menu;
     public final int armorType;
