+++ src/main/java/teamport/aether/mixin/block/CactusSpikeImmunitiesMixin.java
@@ -7,7 +7,7 @@ import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
-import teamport.aether.AetherImmuneMob;
+import teamport.aether.entity.AetherImmuneMob;
 
 @Mixin(value = BlockLogicCactus.class, remap = false)
 public class CactusSpikeImmunitiesMixin {
